remembering the language selection (#7)
Added a custom hook for working with the language.
The language selection is saved to local memory. The next time the page loads, the language has not changed.

Also, the language provider is rendered in separate components.

Now we need to import
DataContext from "../providers/DataContext";

Everything else is compatible.